// enums/HttpStatusCode.ts
export enum HttpStatusCode {
    OK = 200,
    CREATED = 201,
    ACCEPTED = 202, 
    NO_CONTENT = 204, 
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    METHOD_NOT_ALLOWED = 405, 
    CONFLICT = 409, 
    UNPROCESSABLE_ENTITY = 422, 
    INTERNAL_SERVER_ERROR = 500,
    NOT_IMPLEMENTED = 501, 
    SERVICE_UNAVAILABLE = 503, 
    GATEWAY_TIMEOUT = 504,
    }